sensor_monitor/sensors/serializers.py
from rest_framework import serializers
from .models import SensorReading, EndDevice

class SensorReadingSerializer(serializers.ModelSerializer):
    class Meta:
        model = SensorReading
        fields = '__all__'

class EndDeviceSerializer(serializers.ModelSerializer):
    class Meta:
        model = EndDevice
        fields = '__all__'
