requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
dataclasses==0.6
distro==1.9.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
exceptiongroup==1.3.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.6.2
networkx==3.4.2
numpy==2.2.6
ollama==0.5.1
openai==0.28.0
paho-mqtt==2.1.0
pillow==11.2.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==3.12.3
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
